app/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cloudflare==4.3.1
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
segno==1.6.6
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
